+++ src/components/policy/policiesChildren.js
@@ -5,20 +5,17 @@ import PolicyIcon from '@mui/icons-material/Policy';
 import axios from 'axios';
 import { NavLink } from 'react-router-dom';
 import useNotification from '../shared/messages/alerts';
-import { getEnv } from '../../env';
 import * as log from 'loglevel';
 
-const env = getEnv();
-
-export default function PoliciesChildren({ tenantId, tenantName, seTenant, color }) {
-  typeof env.LOG_LEVEL === 'undefined' ? log.setDefaultLevel('debug') : log.setLevel(env.LOG_LEVEL);
-
+export default function PoliciesChildren({ tenantId, tenantName, seTenant, color, env }) {
+  typeof env === 'undefined' ? log.setDefaultLevel('debug') : log.setLevel(env.LOG_LEVEL);
+  const anubisURL = typeof env !== 'undefined' ? env.ANUBIS_API_URL : '';
   // services
   const [msg, sendNotification] = useNotification();
   log.debug(msg);
   const getServices = () => {
     axios
-      .get(env.ANUBIS_API_URL + 'v1/tenants/' + tenantId + '/service_paths')
+      .get(anubisURL + 'v1/tenants/' + tenantId + '/service_paths')
       .then((response) => {
         getPolicies(response.data);
       })
@@ -36,7 +33,7 @@ export default function PoliciesChildren({ tenantId, tenantName, seTenant, color
     let datAccumulator = [];
     for (const service of servicesResponse) {
       axios
-        .get(env.ANUBIS_API_URL + 'v1/policies', {
+        .get(anubisURL + 'v1/policies', {
           headers: {
             'fiware-service': tenantName,
             'fiware-servicepath': service.path