+++ src/components/policy/policiesTable.js
@@ -58,7 +58,7 @@ const DinamicPaper = styled(Paper)(({ theme }) => ({
   }
 }));
 
-export default function PoliciesTable({ data, getData, access_modes, tenantName, agentsTypes, services }) {
+export default function PoliciesTable({ data, getData, access_modes, tenantName, agentsTypes, services, token, env }) {
   // DELETE
   const [openDeleteDialog, setOpenDeleteDialog] = React.useState(false);
 
@@ -437,7 +437,9 @@ export default function PoliciesTable({ data, getData, access_modes, tenantName,
                           {row.resource_type}
                         </TableCell>
                         <TableCell padding="normal" align="left">
-                          {agentToString(row.agent)}
+                          <Typography noWrap gutterBottom sx={{ maxWidth: '70%' }}>
+                            {agentToString(row.agent)}
+                          </Typography>
                         </TableCell>
                         <TableCell padding="normal" align="left">
                           {modeToString(row.mode)}
@@ -482,6 +484,7 @@ export default function PoliciesTable({ data, getData, access_modes, tenantName,
         aria-describedby="edit"
       >
         <PolicyForm
+          env={env}
           tenantName={tenantName}
           action="modify"
           agentsTypes={agentsTypes}
@@ -491,11 +494,13 @@ export default function PoliciesTable({ data, getData, access_modes, tenantName,
           access_modes={access_modes}
           title={<Trans i18nKey="policies.titles.edit" values={{ name: editData.id }} />}
           close={handleClose}
+          token={token}
         ></PolicyForm>
         <DialogActions></DialogActions>
       </DialogRounded>
       <DeleteDialog
         open={openDeleteDialog}
+        env={env}
         onClose={handleCloseDeleteDialog}
         getData={getData}
         data={{