+++ RELEASE_NOTES.md
@@ -4,6 +4,9 @@
 
 ### New features
 
+- The values displayed on the policyForm are from Keycloack
+- Components are now totally indipendent from the API
+
 ### Bug fixes
 
 ### Documentation